fix: hp类型改为有符号，规避无符号相减溢出问题

+++ models/inc/defence.h
@@ -53,7 +53,7 @@ typedef enum {
 
 typedef struct {
     defence_type_t type; /*!< 防具类型 */
-    uint8_t hp;          /*!< 攻击力 */
+    int32_t hp;          /*!< 攻击力 */
     uint8_t weight;      /*!< 重量 */
 } defence_t;
